Move LocalizationMap to `commands` package
