+++ src/main/java/net/dv8tion/jda/api/interactions/commands/LocalizationMap.java
@@ -1,4 +1,4 @@
-package net.dv8tion.jda.api.interactions;
+package net.dv8tion.jda.api.interactions.commands;
 
 import net.dv8tion.jda.api.utils.data.DataObject;
 import net.dv8tion.jda.api.utils.data.SerializableData;
