+++ src/main/java/net/dv8tion/jda/api/interactions/commands/LocalizationMapper.java
@@ -1,6 +1,5 @@
 package net.dv8tion.jda.api.interactions.commands;
 
-import net.dv8tion.jda.api.interactions.LocalizationMap;
 import net.dv8tion.jda.api.interactions.commands.build.CommandData;
 import net.dv8tion.jda.api.interactions.commands.build.SlashCommandData;
 import net.dv8tion.jda.api.utils.data.DataArray;
