+++ src/main/java/net/dv8tion/jda/internal/interactions/CommandDataImpl.java
@@ -16,8 +16,8 @@
 
 package net.dv8tion.jda.internal.interactions;
 
-import net.dv8tion.jda.api.interactions.LocalizationMap;
 import net.dv8tion.jda.api.interactions.commands.Command;
+import net.dv8tion.jda.api.interactions.commands.LocalizationMap;
 import net.dv8tion.jda.api.interactions.commands.LocalizationMapper;
 import net.dv8tion.jda.api.interactions.commands.OptionType;
 import net.dv8tion.jda.api.interactions.commands.build.OptionData;
